Expand SDK capabilities with new DTOs, methods, AI model updates, and enhanced API documentation
- Add `CanUseAi`, `Forms` folder type, and quota-related features (`additional`, `aiSearch`, `DocsCloud` support)
- Enforce `title` as required in `RoomTemplateDto` with length validation
- Replace deprecated `webSearch` with enhanced `search` and `image` capabilities in AI pricing models
- Update data member types (`TotalQuantity` as `long`) and introduce new statistics metrics (`formsUsedSpace`)
- Remove unused private room settings API endpoint, adding HTTP 403 response details

+++ src/DocSpace.API.SDK/Model/ReportDto.cs
@@ -47,7 +47,7 @@ namespace DocSpace.API.SDK.Model
         /// <param name="totalQuantity">The total quantity of operations in the report..</param>
         /// <param name="totalPage">The total number of pages in the report..</param>
         /// <param name="currentPage">The current page number of the report..</param>
-        public ReportDto(List<OperationDto> collection = default, int offset = default, int limit = default, int totalQuantity = default, int totalPage = default, int currentPage = default)
+        public ReportDto(List<OperationDto> collection = default, int offset = default, int limit = default, long totalQuantity = default, int totalPage = default, int currentPage = default)
         {
             this.Collection = collection;
             this.Offset = offset;
@@ -95,7 +95,7 @@ namespace DocSpace.API.SDK.Model
         <example>1</example>
         */
         [DataMember(Name = "totalQuantity", EmitDefaultValue = false)]
-        public int TotalQuantity { get; set; }
+        public long TotalQuantity { get; set; }
 
         /// <summary>
         /// The total number of pages in the report.
